+++ experiments/billable/src/env.ts
@@ -11,6 +11,7 @@ export const envSchema = z.object({
 export type ParsedEnv = z.infer<typeof envSchema>;
 
 export function setupEnv(env: Env) {
+  console.log(env)
   const parsedEnv = envSchema.safeParse(env);
     if (!parsedEnv.success) {
       throw new Error(`Invalid environment variables: ${parsedEnv.error.message}`);
