+++ experiments/billable/wrangler.toml
@@ -5,6 +5,9 @@ compatibility_date = "2024-03-12"
 compatibility_flags = ["nodejs_compat", "rpc"]
 assets = { directory = "./dist/assets/", binding = "ASSETS" }
 
+workers_dev = false
+route = { pattern = "billable.me/*", zone_name = "billable.me" }
+
 [observability]
 enabled = true
